chore: separate `More` views into their own files
