+++ iOS/PadawanWallet/Features/More/MoreView.swift
@@ -84,260 +84,3 @@ struct MoreRootView: View {
         }
     }
 }
-
-struct MoreItem: View {
-    let title: String
-    @Environment(\.padawanColors) private var colors
-
-    var body: some View {
-        HStack {
-            Text(title)
-                .foregroundColor(colors.text)
-            Spacer()
-            Image(systemName: "chevron.right")
-                .foregroundColor(colors.accent2)
-        }
-        .padding()
-        .background(
-            ZStack {
-                // Shadow background
-                RoundedRectangle(cornerRadius: 12)
-                    .fill(.black)
-                    .offset(x: 4, y: 4)
-                
-                // Item background
-                RoundedRectangle(cornerRadius: 12)
-                    .fill(colors.background2)
-            }
-        )
-        .overlay(
-            RoundedRectangle(cornerRadius: 12)
-                .stroke(.black, lineWidth: 2)
-        )
-    }
-}
-
-
-
-struct RecoveryPhraseScreen: View {
-    @Environment(\.padawanColors) private var colors
-    
-    var body: some View {
-        Text("Recovery Phrase Screen")
-            .font(.largeTitle)
-            .foregroundColor(colors.text)
-            .frame(maxWidth: .infinity, maxHeight: .infinity)
-            .background(colors.background)
-            .navigationTitle("Recovery phrase")
-            .navigationBarTitleDisplayMode(.inline)
-    }
-}
-
-struct SendCoinsBackScreen: View {
-    @State private var showCopiedMessage = false
-    @Environment(\.padawanColors) private var colors
-    
-    var body: some View {
-        ScrollView {
-            VStack(spacing: 20) {
-                // QR Code placeholder (since we don't have the vector asset)
-                VStack {
-                    Image(systemName: "qrcode")
-                        .font(.system(size: 200))
-                        .foregroundColor(colors.text)
-                        .background(Color.white)
-                        .cornerRadius(12)
-                        .padding()
-                }
-                .background(colors.background2)
-                .cornerRadius(12)
-                
-                // Bitcoin address with copy functionality
-                VStack(alignment: .leading, spacing: 8) {
-                    Button(action: {
-                        let address = String(localized: "send_coins_back_address")
-                        UIPasteboard.general.string = address
-                        showCopiedMessage = true
-                        DispatchQueue.main.asyncAfter(deadline: .now() + 2) {
-                            showCopiedMessage = false
-                        }
-                    }) {
-                        HStack {
-                            Text(String(localized: "send_coins_back_address"))
-                                .font(.system(size: 14))
-                                .foregroundColor(colors.text)
-                                .multilineTextAlignment(.leading)
-                            
-                            Spacer()
-                            
-                            Image(systemName: showCopiedMessage ? "checkmark.circle.fill" : "doc.on.clipboard")
-                                .foregroundColor(showCopiedMessage ? colors.goGreen : colors.textLight)
-                        }
-                        .padding()
-                        .background(colors.background2)
-                        .cornerRadius(8)
-                    }
-                    
-                    if showCopiedMessage {
-                        Text("Address copied to clipboard!")
-                            .font(.caption)
-                            .foregroundColor(colors.goGreen)
-                            .padding(.horizontal)
-                    }
-                }
-                
-                // Explanatory text
-                Text(String(localized: "send_coins_back"))
-                    .font(.body)
-                    .foregroundColor(colors.textLight)
-                    .multilineTextAlignment(.leading)
-                    .padding(.horizontal)
-                
-                Spacer()
-            }
-            .padding()
-        }
-        .background(colors.background)
-        .navigationTitle(String(localized: "send_signet_coins_back"))
-        .navigationBarTitleDisplayMode(.inline)
-    }
-}
-
-struct LanguageThemeScreen: View {
-    enum Language: String, CaseIterable, Identifiable {
-        case english = "English"
-        case spanish = "Spanish"
-        case portuguese = "Portuguese"
-
-        var id: String { self.rawValue }
-    }
-
-    enum ColorTheme: String, CaseIterable, Identifiable {
-        case tatooine = "Tatooine Desert"
-        case vader = "Vader Dark (Coming Soon!)"
-
-        var id: String { self.rawValue }
-    }
-
-    @State private var selectedLanguage: Language = .english
-    @State private var selectedTheme: ColorTheme = .tatooine
-    @Environment(\.padawanColors) private var colors
-
-    var body: some View {
-        ScrollView {
-            VStack(alignment: .leading, spacing: 32) {
-                // Title + Description
-                VStack(alignment: .leading, spacing: 8) {
-                    Text("Language + Color Theme")
-                        .font(.title2)
-                        .bold()
-                        .foregroundColor(colors.text)
-
-                    Text("Select your preferred language and color theme for Padawan Wallet.")
-                        .foregroundColor(colors.textLight)
-                        .font(.body)
-                }
-
-                // Language picker
-                VStack(alignment: .leading, spacing: 16) {
-                    Text("App-Level Language")
-                        .bold()
-                        .font(.headline)
-                        .foregroundColor(colors.text)
-                        .padding(.bottom, 4)
-
-                    Divider()
-                        .background(colors.textFaded)
-
-                    ForEach(Language.allCases) { language in
-                        HStack {
-                            Image(systemName: selectedLanguage == language ? "circle.inset.filled" : "circle")
-                                .foregroundColor(selectedLanguage == language ? colors.accent1 : colors.textLight)
-                            Text(language.rawValue)
-                                .foregroundColor(colors.text)
-                            Spacer()
-                        }
-                        .contentShape(Rectangle())
-                        .onTapGesture {
-                            selectedLanguage = language
-                        }
-                    }
-                }
-
-                // Color theme picker
-                VStack(alignment: .leading, spacing: 16) {
-                    Text("Color Theme")
-                        .bold()
-                        .font(.headline)
-                        .foregroundColor(colors.text)
-                        .padding(.bottom, 4)
-
-                    Divider()
-                        .background(colors.textFaded)
-
-                    ForEach(ColorTheme.allCases) { theme in
-                        HStack {
-                            Image(systemName: selectedTheme == theme ? "circle.inset.filled" : "circle")
-                                .foregroundColor(theme == .vader ? colors.textFaded : colors.accent1)
-                            Text(theme.rawValue)
-                                .foregroundColor(theme == .vader ? colors.textFaded : colors.text)
-                            Spacer()
-                        }
-                        .contentShape(Rectangle())
-                        .onTapGesture {
-                            if theme != .vader {
-                                selectedTheme = theme
-                            }
-                        }
-                    }
-                }
-
-                Spacer()
-            }
-            .padding()
-        }
-        .background(colors.background)
-        .navigationTitle("")
-        .navigationBarTitleDisplayMode(.inline)
-    }
-}
-
-struct AboutPadawanScreen: View {
-    @Environment(\.padawanColors) private var colors
-    
-    var body: some View {
-        ScrollView {
-            VStack(alignment: .leading, spacing: 24) {
-                // About text
-                Text(String(localized: "about_text"))
-                    .font(.body)
-                    .foregroundColor(colors.textLight)
-                    .multilineTextAlignment(.leading)
-                
-                // Privacy text
-                Text(String(localized: "privacy_text"))
-                    .font(.body)
-                    .foregroundColor(colors.textLight)
-                    .multilineTextAlignment(.leading)
-                
-                // Privacy policy link
-                Button(action: {
-                    if let url = URL(string: String(localized: "privacy_link")) {
-                        UIApplication.shared.open(url)
-                    }
-                }) {
-                    Text("Read our privacy policy here.")
-                        .font(.body)
-                        .foregroundColor(colors.accent2)
-                        .underline()
-                }
-                
-                Spacer()
-            }
-            .padding()
-        }
-        .background(colors.background)
-        .navigationTitle(String(localized: "about_padawan"))
-        .navigationBarTitleDisplayMode(.inline)
-    }
-}
